Final Code
You will need to download Selenium Urlib as well as add an __init__.py as it does let me upload it. As well You will need to download the correct webdriver for you chrome browser, finally you will need to add your username and password to the config file. I have added my details on canvas if you need to run it but can't and need to reach me.
Thank you again professor